+++ packages/server/graphql/public/mutations/acceptRequestToJoinDomain.ts
@@ -126,13 +126,8 @@ const acceptRequestToJoinDomain: MutationResolvers['acceptRequestToJoinDomain'] 
   await redisLock.unlock()
 
   // Send the new team member a welcome & a new token
-  const updatedUser = await getUserById(createdBy)
-  if (!updatedUser) {
-    return standardError(new Error('User not found'))
-  }
-  publish(SubscriptionChannel.NOTIFICATION, userId, 'AuthTokenPayload', {
-    tms: updatedUser.tms
-  })
+  const tms = await dataLoader.get('teamIdsByUserId').load(userId)
+  publish(SubscriptionChannel.NOTIFICATION, userId, 'AuthTokenPayload', {tms})
 
   validTeams.forEach((team) => {
     const {id: teamId} = team