+++ packages/server/postgres/types/NewMeetingPhase.d.ts
@@ -174,6 +174,7 @@ export interface UpdatesPhase extends GenericMeetingPhase {
 export type RetroMeetingPhase =
   | CheckInPhase
   | TeamHealthPhase
+  | UpdatesPhase
   | ReflectPhase
   | GroupPhase
   | VotePhase